% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{BatchDownload}
\alias{BatchDownload}
\title{Description: Base functions and variables.}
\usage{
BatchDownload(links, outdir, method = "curl", pbar = TRUE)
}
\arguments{
\item{links}{One-dimensional array containing urls.}

\item{outdir}{The output folder.}

\item{method}{Url downloading method, default is \code{curl}. It can be
"internal", "wininet" (Windows only) "libcurl", "wget" and "curl".}

\item{pbar}{Indicate whether to show a progress bar. Default is \code{TRUE}.}
}
\value{
\code{NULL} if successful, otherwise will throw out errors.
}
\description{
This file process batch downloading from a bunch of provided links. It can
be used to download both the image query result from the Appears application
and general files. The input of this script is a "*.txt" file that contains
downloadable links.
\code{links} must be a one-dimensional array containing urls needed to be
downloaded as array rows. If the specified \code{outdir} does not exist, the
function will create it.
}
\details{
Author: Xiaojie Gao
Date: 2023-02-06

Batch download links into a specified folder
}
