% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usgs_m2m_api.R
\name{DatasetClearCustomization}
\alias{DatasetClearCustomization}
\title{I don't understand what this function is for}
\usage{
DatasetClearCustomization(token, dataset_name = "", metadata_type = c(""))
}
\arguments{
\item{token}{The API key.}

\item{dataset_name}{Used to identify the dataset to clear. If null, all dataset
customizations will be cleared.}

\item{metadataType}{If populated, identifies which metadata to clear (export,
full, res_sum, shp)}
}
\description{
I don't understand what this function is for
}
