% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_gimmis_ndvi.R
\name{DlGimmisNDVI}
\alias{DlGimmisNDVI}
\title{Description: Download the global GIMMS NDVI3g product.}
\usage{
DlGimmisNDVI(outdir)
}
\arguments{
\item{outdir}{Output directory.}
}
\value{
\code{NULL}
}
\description{
From : https://data.tpdc.ac.cn/en/data/c6113f70-884a-4716-98e3-933421c57f25/?q=gimms
Ftp server: 210.72.14.198
Ftp username: download_414169
Ftp password: 31879907
}
\details{
Author: Xiaojie Gao
Date: 2022-09-22

Download the global GIMMS NDVI3g product
}
