% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{FtpQuery}
\alias{FtpQuery}
\title{Query FTP folder}
\usage{
FtpQuery(ftp_url, username, password, dirlistonly = TRUE)
}
\arguments{
\item{ftp_url}{FTP address.}

\item{username, password}{FTP username and password.}

\item{dirlistonly}{Whether to return directories only.}
}
\value{
Query result in a table.
}
\description{
Query files and directories in an FTP server folder.
}
