% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usgs_m2m_api.R
\name{GetBulkProducts}
\alias{GetBulkProducts}
\title{Lists all available bulk products for a dataset - this does not guarantee
scene availability.}
\usage{
GetBulkProducts(token, dataset_name)
}
\arguments{
\item{token}{The API key}

\item{dataset_name}{Used to identify which dataset to return results for}
}
\value{
All available bulk products for a dataset
}
\description{
Lists all available bulk products for a dataset - this does not guarantee
scene availability.
}
