% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usgs_m2m_api.R
\name{GetDataOwner}
\alias{GetDataOwner}
\title{Get the contact information of the data owner.}
\usage{
GetDataOwner(token, data_owner)
}
\arguments{
\item{token}{The API key.}

\item{data_owner}{Used to identify the data owner - this value comes from the
dataset-search response.}
}
\value{
The contact information list.
}
\description{
Get the contact information of the data owner.
}
