% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usgs_m2m_api.R
\name{GetDataset}
\alias{GetDataset}
\title{Search a data product by id or name}
\usage{
GetDataset(token, dataset_id = "", dataset_name = "")
}
\arguments{
\item{token}{The API key}

\item{dataset_id}{The dataset identifier - must use this or datasetName}

\item{dataset_name}{The system-friendly dataset name - must use this or datasetId}
}
\value{
The dataset information object
}
\description{
Search a data product by id or name
}
