% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usgs_m2m_api.R
\name{GetDatasetCoverage}
\alias{GetDatasetCoverage}
\title{Returns coverage for a given dataset.}
\usage{
GetDatasetCoverage(token, dataset_name = "")
}
\arguments{
\item{token}{The API key.}

\item{dataset_name}{Determines which dataset to return coverage for.}
}
\value{
A list containing the coverage bounding box \code{bbox} and actual
polygon(s) \code{shp}. Both \code{bbox} and \code{shp} are \code{SpatialPolygonsDataFrame} objects.
}
\description{
Returns coverage for a given dataset.
}
