% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usgs_m2m_api.R
\name{GetDatasetCustomization}
\alias{GetDatasetCustomization}
\title{This method is used to retrieve metadata customization for a specific dataset.}
\usage{
GetDatasetCustomization(token, dataset_name)
}
\arguments{
\item{dataset_name}{Determines which dataset to return filters for}
}
\description{
This method is used to retrieve metadata customization for a specific dataset.
}
