% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usgs_m2m_api.R
\name{GetDatasetDownloadOptions}
\alias{GetDatasetDownloadOptions}
\title{This request lists all available products for a given dataset - this does not
guarantee scene availability.}
\usage{
GetDatasetDownloadOptions(token, dataset_name = "", scene_filter = NULL)
}
\arguments{
\item{token}{The API key.}

\item{dataset_name}{Used to identify the which dataset to return results for}

\item{scene_filter}{Used to filter data within the dataset.
See \href{https://m2m.cr.usgs.gov/api/docs/datatypes/#sceneFilter}{here}}
}
\value{
A list containing all available datasets for download
}
\description{
This request lists all available products for a given dataset - this does not
guarantee scene availability.
}
