% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usgs_m2m_api.R
\name{GetDatasetFilters}
\alias{GetDatasetFilters}
\title{This request is used to return the metadata filter fields for the
specified dataset. These values can be used as additional criteria when
submitting search and hit queries.}
\usage{
GetDatasetFilters(token, dataset_name)
}
\arguments{
\item{token}{The API key.}

\item{dataset_name}{Determines which dataset to return filters for}
}
\value{
Metadata filter fields
}
\description{
This request is used to return the metadata filter fields for the
specified dataset. These values can be used as additional criteria when
submitting search and hit queries.
}
