% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usgs_m2m_api.R
\name{GetDatasetMetadata}
\alias{GetDatasetMetadata}
\title{This method is used to retrieve all metadata fields for a given dataset.}
\usage{
GetDatasetMetadata(token, dataset_name)
}
\arguments{
\item{token}{The API key.}

\item{dataset_name}{Determines which dataset to return filters for}
}
\value{
Metadata information
}
\description{
This method is used to retrieve all metadata fields for a given dataset.
}
