% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usgs_m2m_api.R
\name{GetDatasetOrderProducts}
\alias{GetDatasetOrderProducts}
\title{Lists all available order products for a dataset - this does not guarantee
scene availability.}
\usage{
GetDatasetOrderProducts(token, dataset_name)
}
\arguments{
\item{token}{The API key.}

\item{dataset_name}{Determines which dataset to return filters for}
}
\value{
Order product information
}
\description{
Lists all available order products for a dataset - this does not guarantee
scene availability.
}
