% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usgs_m2m_api.R
\name{Login}
\alias{Login}
\title{Login to get the API key.
Need to register an ERS account at: https://m2m.cr.usgs.gov/}
\usage{
Login(ers_username, ers_password)
}
\arguments{
\item{ers_username}{ERS username}

\item{ers_password}{ERS password}
}
\value{
The API key
}
\description{
Login to get the API key.
Need to register an ERS account at: https://m2m.cr.usgs.gov/
}
