% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usgs_m2m_api.R
\name{SearchDatasetsInCatalog}
\alias{SearchDatasetsInCatalog}
\title{List all the datasets available under a category.}
\usage{
SearchDatasetsInCatalog(
  token,
  catalog = "",
  include_msg = NULL,
  public_only = NULL,
  use_customization = NULL,
  parent_id = "",
  dataset_filter = ""
)
}
\arguments{
\item{token}{The API key returned by the \code{Login} function}

\item{catalog}{Used to identify datasets that are associated with a given
application}

\item{include_msg}{Optional parameter to include messages regarding specific
dataset components}

\item{public_only}{Used as a filter out datasets that are not accessible to
unauthenticated general public users}

\item{use_customization}{Used as a filter out datasets that are excluded by
user customization}

\item{parent_id}{If provided, returned categories are limited to categories that
are children of the provided ID}

\item{dataset_filter}{If provided, filters the datasets - this automatically adds
a wildcard before and after the input value}
}
\value{
The dataset information
}
\description{
List all the datasets available under a category.
}
