% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usgs_m2m_api.R
\name{SendRequest}
\alias{SendRequest}
\title{Internal function used to send requests to the server}
\usage{
SendRequest(token, path, params = "")
}
\arguments{
\item{token}{The API key}

\item{path}{The API path}

\item{params}{Other API needed parameters}
}
\value{
The response object by the server
}
\description{
Internal function used to send requests to the server
}
